# tests/Swagger_updated_20-10-2017.yaml
swagger: '2.0'
info:
  version: '2.0'
  title: "API Project I'm Ready"
  description: "API Project I'm Ready"
# basePath: "/"
# host: "imready.ml"
...
produces:
  - "application/json"
schemes:
 - "https"
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "Authorization"
    in: "header"
...
definitions:
...
  Caregivers:
    type: "object"
    properties:
      id:
        type: "string"
      name:
        type: "string"
        description: "Name of the user"
      role:
        description: "Role of client"
        type: "string"
        enum:
        - "admin"
        - "client"
        - "caregiver"
        - "family"
      messages:
        description: "All messages this user has sent or received"
        type: "array"
        items: {
          $ref: "#/definitions/Messages"
        }        
      appointments:
        description: "All appointments of the user"
        type: "array"
        items: {
          $ref: "#/definitions/Appointments"
        }
  EntityModel:
    type: "object"
    properties:
      id:
        type: "string"
...
  Appointments:
    type: "object"
    properties:
      id:
        description: "ID of the appointment"
        type: "string"
      datetime:
        description: "Deadline of the appointment"
        type: "string"
        format: "datetime"
      description:
        description: "Description of the appointment"
        type: "string"    
...
  Messages:
    type: "object"
    properties:
      id:
        description: "ID of the message"
        type: "string"
      senderId:
        description: "ID of the sender of the message"
        type: "string"
      receiverId:
        description: "ID of the receiver of the message"
        type: "string"
      message:
        description: "Content of the message"
        type: "string"
      datetime:
        description: "Time the message was sent"
        type: "string"
        format: "datetime"
      read:
        description: "The read status of the message"
        type: "boolean"
...
  Credentials:
    type: "object"
    properties:
      username: 
        description: "Username"
        type: "string"
      password:
        description: "Password"
        type: "string"
...
  ErrorResponses:
    type: "object"
    properties:
      code:
        description: "Error code"
        type: "integer"
        format: "int64"
      message:
        description: "Error message"
        type: "string"        
paths:
...
  /caregivers/{id}:
    get:
      summary: "Get caregiver"
      description: "Returns a caregiver that fits the specified arguments."
      tags:
      - "general"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of caregiver to return"
        required: true
        type: "string"
      responses:
        200:
          description: "succesful operation"
          schema:
            $ref: "#/definitions/Caregivers"
        400:
          description: "Invalid parameters"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "User not authorized"
          schema:
            $ref: "#/definitions/ErrorResponses"
        404:
          description: "Caregiver not found"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []

    put:
      summary: "Update caregiver"
      description: "Update caregiver with specified ID"
      tags:
      - "general"
      consumes: 
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of caregiver to be updated"
        required: true
        type: "string"
      - name: "Credentials"
        in: "body"
        description: "Credentials of the user"
        required: true
        schema:
          $ref: "#/definitions/Credentials"
      responses:
        200:
          description: "succesfully updated caregiver"
        400:
          description: "Invalid ID supplied"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "User not authorized"
          schema:
            $ref: "#/definitions/ErrorResponses"
        404:
          description: "Caregiver not found"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []

    delete:
      summary: "Delete caregiver"
      description: "Delete caregiver with specified ID"
      tags:
      - "general"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of the caregiver to deleted"
        required: true
        type: "string"
      responses:
        204:
          description: "succesfully deleted caregiver"
        400:
          description: "Invalid ID supplied"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "User not authorized"
          schema:
            $ref: "#/definitions/ErrorResponses"
        404:
          description: "Caregiver not found"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []
...
  /clients/{clientId}/caregivers/{caregiverId}:
    get:
      summary: "Get a single client's caregiver with ID"
      description: "Returns a single caregiver with specified userID and caregiverID"
      tags:
        - "connections"
      parameters:
      - name: "clientId"
        in: "path"
        description: "ID of client to be returned"
        required: true
        type: "string"
      - name: "caregiverId"
        in: "path"
        description: "ID of the client's caregiver to be returned"
        required: true
        type: "string"
      responses:
        200:
          description: "succesfully found the caregigver by ID"
          schema:
            $ref: "#/definitions/Caregivers"
        400:
          description: "Invalid ID supplied"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "User not authorized"
          schema:
            $ref: "#/definitions/ErrorResponses"
        404:
          description: "Client or caregiver not found"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []

    put:
      summary: "Update client's caregiver"
      description: "Update caregiver with specified userID and caregiverID"
      tags:
      - "connections"
      consumes:
      - "application/json"
      parameters:
      - name: "clientId"
        in: "path"
        description: "ID of client to be updated"
        required: true
        type: "string"
      - name: "caregiverId"
        in: "path"
        description: "ID of the client's caregiver to be updated"
        required: true
        type: "string"
      - name: "caregiver"
        in: "body"
        description: "Updated caregiver object"
        required: true
        schema:
          $ref: "#/definitions/EntityModel"
      responses:
        200:
          description: "Succesfully updated client"
        400:
          description: "Invalid ID supplied"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "User not authorized"
          schema:
            $ref: "#/definitions/ErrorResponses"
        404:
          description: "Client or caregiver not found"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []

    delete:
      summary: "Delete client's caregiver"
      description: "Delete client's caregiver with specified ID"
      tags:
      - "connections"
      parameters:
      - name: "clientId"
        in: "path"
        description: "ID of client to be updated"
        required: true
        type: "string"
      - name: "caregiverId"
        in: "path"
        description: "ID of the client's caregiver to be deleted"
        required: true
        type: "string"
      responses:
        204:
          description: "Succesfully updated client"
        400:
          description: "Invalid ID supplied"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "User not authorized"
          schema:
            $ref: "#/definitions/ErrorResponses"
        404:
          description: "Client not found"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []
...
  /clients/{clientId}/appointments/{appointmentId}:
    get:
      summary: "Get the client specific appointment."
      description: "Returns an appointments that is specific and tied to the client."
      tags:
      - "appointments"
      parameters:
      - name: "clientId"
        in: "path"
        required: true
        type: "string"
        description: "Needs the clients Id to identify."
      - name: "appointmentId"
        in: "path"
        required: true
        type: "string"
        description: "Needs an Id to identify an appointment."
      responses:
        200:
          description: "Client specific appointment is succesfully returned"
          schema:
            $ref: "#/definitions/Appointments"
        400:
          description: "The input parameter(s) is invalid"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "Unauthorized request"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []

    put:
      summary: "Update the clients appointment"
      description: "update a specfic appointment"
      tags:
      - "appointments"
      consumes:
      - "application/json"
      parameters:
      - name: "clientId"
        in: "path"
        description: "Needs a client ID to link the appointment to"
        required: true
        type: "string"
      - name: "appointmentId"
        in: "path"
        description: "Needs a Id to update the appointment."
        required: true
        type: "string"
      - name: "appointment"
        in: "body"
        description: "appointment object with ISO DateTime format to be replaced"
        required: true
        schema:
          $ref: "#/definitions/Appointments"
      responses:
        200:
          description: "The appointment has been succesfully updated"
        400:
          description: "The input parameter(s) is invalid"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "An unauthorized attempt"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []

    delete:
      summary: "The clients appointment can be deleted"
      description: "Deletes a clients appointment"
      tags:
      - "appointments"
      parameters:
      - name: "clientId"
        in: "path"
        required: true
        description: "Needs a client ID to link the appointment to"
        type: "string"
      - name: "appointmentId"
        in: "path"
        description: "Needs a Id to update the appointment."
        required: true
        type: "string"
      responses:
        204:
          description: "The appointment has been succesfully deleted"
        400:
          description: "The input parameter(s) is invalid"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "An unauthorized attempt"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []
  /caregivers/{caregiverId}/appointments:
    get:
      summary: "Get all the caregiver's appointments."
      description: "Returns all appointments that are specific and tied to the caregiver."
      tags:
      - "appointments"
      parameters:
      - name: "caregiverId"
        in: "path"
        required: true
        type: "string"
        description: "Needs an Id to identify the caregiver."
      - name: "count"
        in: "query"
        description: "The maximum number of appointments to return."
        required: false
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "All caregiver's appointments are succesfully returned"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Appointments"
        400:
          description: "The input parameter is invalid"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "An unauthorized attempt"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []

    post:
      summary: "Add a new appointment for the caregiver"
      description: "Add an appointment"
      tags:
      - "appointments"
      consumes:
      - "application/json"
      parameters:
      - name: "caregiverId"
        in: "path"
        description: "Needs a caregiver ID to link the appointment to"
        required: true
        type: "string"
      - name: "appointment"
        in: "body"
        description: "Appointment object with ISO DateTime format"
        required: true
        schema:
          $ref: "#/definitions/Appointments"
      responses:
        201:
          description: "The appointment has been succesfully added"
        400:
          description: "The input parameter(s) are invalid"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "An unauthorized attempt"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []
  /caregivers/{caregiverId}/appointments/{appointmentId}:
    get:
      summary: "Get the caregiver specific appointment."
      description: "Returns an appointment that is specific and tied to the caregiver."
      tags:
      - "appointments"
      parameters:
      - name: "caregiverId"
        in: "path"
        required: true
        description: "Needs a caregiver ID to link the appointment to"
        type: "string"
      - name: "appointmentId"
        in: "path"
        description: "Needs a Id to update the appointment."
        required: true
        type: "string"
      responses:
        200:
          description: "Caregiver's specific appointment is succesfully returned"
          schema:
            $ref: "#/definitions/Appointments"
        400:
          description: "The input parameter(s) is invalid"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "Unauthorized request"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []

    put:
      summary: "Update the caregiver's appointment"
      description: "Update a specfic appointment"
      tags:
      - "appointments"
      consumes:
      - "application/json"
      parameters:
      - name: "caregiverId"
        in: "path"
        required: true
        description: "Needs a caregiver ID to link the appointment to"
        type: "string"
      - name: "appointmentId"
        in: "path"
        description: "Needs a Id to update the appointment."
        required: true
        type: "string"
      - name: "appointment"
        in: "body"
        description: "appointment object with ISO DateTime to be replaced"
        required: true
        schema:
          $ref: "#/definitions/Appointments"
      responses:
        200:
          description: "The appointment has been succesfully updated"
        400:
          description: "The input parameter(s) is invalid"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "An unauthorized attempt"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []

    delete:
      summary: "The caregiver's appointment can be deleted"
      description: "Deletes a caregiver's appointment"
      tags:
      - "appointments"
      parameters:
      - name: "caregiverId"
        in: "path"
        required: true
        description: "Needs a caregiver ID to link the appointment to"
        type: "string"
      - name: "appointmentId"
        in: "path"
        description: "Needs a Id to update the appointment."
        required: true
        type: "string"
      responses:
        204:
          description: "The appointment has been succesfully deleted"
        400:
          description: "The input parameter(s) is invalid"
          schema:
            $ref: "#/definitions/ErrorResponses"
        401:
          description: "An unauthorized attempt"
          schema:
            $ref: "#/definitions/ErrorResponses"
      security:
      - api_key: []
...
host: virtserver.swaggerhub.com
basePath: /InhollandCloudAPI/ProjectAPI/2.0